Added feed controller, moved all feeds logic to it
