+++ app/assets/stylesheets/feed.scss
@@ -1,13 +1,5 @@
 .feed {
-  .status_arrow {
-    top: 4px
-  }
-
-  .unread_event {
-    margin-right: 5px;
-  }
-
-  .post.unread{
-    background-color: rgb(247, 247, 247);
+  .post.unread {
+    background-color: #e8e7ff;
   }
 }
