+++ app/controllers/users_controller.rb
@@ -1,18 +1,10 @@
 class UsersController < ApplicationController
-  include Feed
   before_action :find_user, only: [:show]
-  after_action :mark_as_read_loaded, only: [:feed]
   skip_before_action :require_active_user, only: [:show, :ranking]
   skip_before_action :authenticate_user!, only: [:ranking]
 
   def show
     issue_listing(@user.issues)
-  end
-
-  def feed
-    current_user.check_notifications unless request.xhr?
-    @events = current_user.events.ordered.page(params[:page]).per(3)
-    respond_for_feed(@events)
   end
 
   def ranking
@@ -26,11 +18,5 @@ class UsersController < ApplicationController
 
   def find_user
     @user = User.find(params[:id])
-  end
-
-  def mark_as_read_loaded
-    return unless request.xhr?
-    notifications = current_user.unread_notifications_for(@events)
-    notifications.update_all(readed: true)
   end
 end
