+++ app/helpers/feeds_helper.rb
@@ -1,0 +1,5 @@
+module FeedsHelper
+  def users_feed?
+    params[:controller] == 'feeds' && params[:action] == 'user_feed'
+  end
+end
