--- app/helpers/notifications_helper.rb
@@ -1,6 +1,0 @@
-# frozen_string_literal: true
-module NotificationsHelper
-  def users_feed?
-    params[:controller] == 'users'
-  end
-end
