+++ spec/controllers/feeds_controller_spec.rb
@@ -1,0 +1,4 @@
+require 'rails_helper'
+
+RSpec.describe FeedsController, type: :controller do
+end
